# common/swagger.json
{"openapi":"3.0.0","paths":{"/api/products/populate":{"post":{"operationId":"ProductsController_populate","summary":"Populates the DB with random values","parameters":[],"responses":{"201":{"description":""}},"tags":["products"]}},"/api/products":{"post":{"operationId":"ProductsController_create","summary":"Creates a product","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductsDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductsDto"}}}}},"tags":["products"]},"get":{"operationId":"ProductsController_getAll","summary":"Get all products in the store","parameters":[{"name":"query","required":false,"in":"query","description":"FTS query text","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","description":"Products statuses separated by comma. Allowed values: Limited,New","example":"New,Limited","schema":{"type":"string"}},{"name":"category","required":false,"in":"query","description":"Product category. Allowed values: Canola,Corn,Oats,Wheat,Soybeans,Barley","example":"Barley","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductsDto"}}}}}},"tags":["products"]},"put":{"operationId":"ProductsController_update","summary":"Update product info","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductsDto"}}}},"responses":{"200":{"description":""}},"tags":["products"]},"delete":{"operationId":"ProductsController_deleteAll","parameters":[],"responses":{"200":{"description":""}},"tags":["products"]}},"/api/products/{id}":{"get":{"operationId":"ProductsController_getOne","summary":"Get a single product","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductsDto"}}}}},"tags":["products"]},"delete":{"operationId":"ProductsController_delete","summary":"Delete a product","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["products"]}},"/api/products/{id}/image":{"post":{"operationId":"ProductsController_upload","summary":"Upload an image to the server's disk storage","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductsDto"}}}}},"tags":["products"]}}},"info":{"title":"Store API","description":"Store REST API documentation.","version":"v0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ProductsDto":{"type":"object","properties":{"name":{"type":"string","example":"SC3000","description":"Product display name"},"description":{"type":"string","example":"An amazing product with unique features","description":"Product display description"},"price":{"type":"number","example":"40","description":"Product price in USD"},"image":{"type":"string","example":"https://ima.ge/product.png","description":"Product thumbnail"},"discount":{"type":"string","example":"5","description":"USD discount amount"},"status":{"type":"array","example":["Limited","New"],"description":"Status array","items":{"type":"string","enum":["Limited","New"]}},"category":{"type":"string","example":"Canola","description":"Product category","enum":["Canola","Corn","Oats","Wheat","Soybeans","Barley"]}},"required":["name","description","price","image","discount","status","category"]}}}}